Rename EditDeck and move to pages folder
