+++ src/client/App.tsx
@@ -5,13 +5,13 @@ import PrivateRoute from './components/routing/PrivateRoute';
 import PublicRoute from './components/routing/PublicRoute';
 import { useAuth } from './contexts/AuthContext';
 import { useTheme } from './contexts/ThemeContext';
-import EditDeckLayout from './layouts/EditDeckLayout';
 import AddCard from './pages/AddCard';
 import DeckCards from './pages/DeckCards';
 import Decks from './pages/Decks';
 import DeckSettings from './pages/DeckSettings';
 import DeckTags from './pages/DeckTags';
 import EditCard from './pages/EditCard';
+import EditDeck from './pages/EditDeck';
 import HomeLoggedIn from './pages/HomeLoggedIn';
 import HomeLoggedOut from './pages/HomeLoggedOut';
 import Login from './pages/Login';
@@ -34,7 +34,7 @@ export default function App() {
           <Route path='/' element={authed ? <HomeLoggedIn /> : <HomeLoggedOut />} />
           <Route element={<PrivateRoute />}>
             <Route path='/decks' element={<Decks />} />
-            <Route path='/decks/:deckId' element={<EditDeckLayout />}>
+            <Route path='/decks/:deckId' element={<EditDeck />}>
               <Route index element={<Navigate to='cards' replace />} />
               <Route path='cards' element={<DeckCards />} />
               <Route path='cards/new' element={<AddCard />} />
